Fix hibernate issues causing crash against 01.00.00 liquibase package.

+++ backend/src/main/java/ca/bc/gov/nrs/vdyp/backend/data/models/CalculationEngineCodeModel.java
@@ -1,5 +1,6 @@
 package ca.bc.gov.nrs.vdyp.backend.data.models;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import lombok.Data;
@@ -8,7 +9,7 @@ import lombok.Data;
 public class CalculationEngineCodeModel {
 	private String calculationEngineCode;
 	private String description;
-	private Integer displayOrder;
+	private BigDecimal displayOrder;
 	private Date effectiveDate;
 	private Date expiryDate;
